yay! Got the canvas working with a blinking animation!, working on actual drawing now.
